ktesting: reorganize implementation code
The goal is to make it easier to read the source top-to-bottom.
The previous organization had historic reasons.

The structure is now:
- doc.go: package docs
- tcontext.go: TContext and most of its methods
- assert.go: Gomega integration
- signal*.go: signal handling, including progress reporting
- contexthelper.go: utility code
- ktesting.go: global setup and compatibility with klog's ktesting

This was done purely via copy-and-paste, with one exception: run(tCtx, ...)
became tCtx.run and the comment was dropped because it was obvious.

Tests were not touched and still pass.
